Camada Controls: Adicionado a classe SellProductMethods, para que a mesma receba os métodos de seu respectivo form.
